implement horizontal-scalability app

+++ communication-support/frontend/scripts/stop.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 abs_path=$(cd $(dirname $0)/.. && pwd)
 
-pid=$(cat $abs_path/pid)
-kill -9 $pid
-
-rm $abs_path/pid
+if [ -f $abs_path/pid ];then
+  pid=$(cat $abs_path/pid)
+  kill -9 $pid
+  rm $abs_path/pid
+  echo 'stopped!'
+fi
